moves isRequestGroup
