+++ packages/insomnia-app/app/common/export.ts
@@ -23,11 +23,11 @@ import { trackEvent } from './analytics';
 import {
   isProtoDirectory,
   isProtoFile,
-  isRequestGroup,
   isWorkspace,
 } from '../models/helpers/is-model';
 import { isGrpcRequest } from '../models/grpc-request';
 import { isRequest } from '../models/request';
+import { isRequestGroup } from '../models/request-group';
 
 const EXPORT_FORMAT = 4;
