+++ packages/insomnia-app/app/common/sorting.ts
@@ -11,8 +11,7 @@ import {
 } from './constants';
 import { isRequest, Request } from '../models/request';
 import { GrpcRequest, isGrpcRequest } from '../models/grpc-request';
-import type { RequestGroup } from '../models/request-group';
-import { isRequestGroup } from '../models/helpers/is-model';
+import { isRequestGroup, RequestGroup } from '../models/request-group';
 
 type SortableModel = Request | RequestGroup | GrpcRequest;
 type SortFunction = (a: SortableModel, b: SortableModel) => number;
