+++ packages/insomnia-app/app/ui/components/templating/tag-editor.tsx
@@ -18,8 +18,7 @@ import { metaSortKeySort } from '../../../common/sorting';
 import type { BaseModel } from '../../../models/index';
 import type { Workspace } from '../../../models/workspace';
 import { isRequest, Request } from '../../../models/request';
-import type { RequestGroup } from '../../../models/request-group';
-import { isRequestGroup } from '../../../models/helpers/is-model';
+import { isRequestGroup, RequestGroup } from '../../../models/request-group';
 import type { PluginArgumentEnumOption } from '../../../templating/extensions/index';
 import { Dropdown, DropdownButton, DropdownDivider, DropdownItem } from '../base/dropdown/index';
 import FileInputButton from '../base/file-input-button';
